requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
# Editable install with no version control (clip==1.0)
-e /home/dyh/project/AICity2022Track2
filelock==3.6.0
ftfy==6.1.1
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
timm==0.5.4
tokenizers==0.11.6
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.63.0
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
yacs==0.1.8
